requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
--index-url https://bbpteam.epfl.ch/repository/devpi/simple

anyio==3.7.0
    # via
    #   starlette
    #   watchfiles
asttokens==2.2.1
    # via devtools
attrs==23.1.0
    # via
    #   entity-management
    #   jsonschema
cached-property==1.5.2
    # via tmd
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   diameter-synthesis
    #   neurom
    #   uvicorn
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
deprecation==2.1.0
    # via python-keycloak
devtools[pygments]==0.11.0
    # via entity-management
diameter-synthesis==0.5.4
    # via neurots
ecdsa==0.18.0
    # via python-jose
entity-management==1.2.19
    # via morphology-synthesis (pyproject.toml)
exceptiongroup==1.1.1
    # via anyio
executing==1.2.0
    # via devtools
fastapi==0.98.0
    # via morphology-synthesis (pyproject.toml)
fonttools==4.40.0
    # via matplotlib
h11==0.14.0
    # via uvicorn
h5py==3.9.0
    # via tmd
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
isodate==0.6.1
    # via rdflib
joblib==1.3.0
    # via scikit-learn
jsonschema==4.17.3
    # via
    #   diameter-synthesis
    #   neurots
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.7.1
    # via
    #   diameter-synthesis
    #   neurom
    #   neurots
morphio==3.3.5
    # via
    #   diameter-synthesis
    #   neurom
    #   neurots
    #   tmd
munkres==1.1.4
    # via tmd
neurom==3.2.2
    # via
    #   diameter-synthesis
    #   neurots
neurots==3.4.0
    # via morphology-synthesis (pyproject.toml)
numpy==1.25.0
    # via
    #   contourpy
    #   diameter-synthesis
    #   h5py
    #   matplotlib
    #   morphio
    #   neurom
    #   neurots
    #   pandas
    #   scikit-learn
    #   scipy
    #   tmd
packaging==23.1
    # via
    #   deprecation
    #   matplotlib
pandas==2.0.3
    # via
    #   diameter-synthesis
    #   neurom
pillow==9.5.0
    # via matplotlib
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pydantic==1.10.9
    # via
    #   fastapi
    #   morphology-synthesis (pyproject.toml)
pygments==2.15.1
    # via devtools
pyjwt==2.7.0
    # via entity-management
pyparsing==3.1.0
    # via
    #   matplotlib
    #   rdflib
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   entity-management
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via uvicorn
python-jose==3.3.0
    # via python-keycloak
python-keycloak==3.3.0
    # via entity-management
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   morphology-synthesis (pyproject.toml)
    #   neurom
    #   uvicorn
rdflib==6.3.2
    # via
    #   entity-management
    #   sparqlwrapper
requests==2.31.0
    # via
    #   entity-management
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rsa==4.9
    # via python-jose
scikit-learn==1.2.2
    # via tmd
scipy==1.11.1
    # via
    #   diameter-synthesis
    #   neurom
    #   neurots
    #   scikit-learn
    #   tmd
six==1.16.0
    # via
    #   asttokens
    #   ecdsa
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sparqlwrapper==2.0.0
    # via entity-management
starlette==0.27.0
    # via fastapi
threadpoolctl==3.1.0
    # via scikit-learn
tmd==2.3.1
    # via neurots
tqdm==4.65.0
    # via neurom
typing-extensions==4.7.0
    # via pydantic
tzdata==2023.3
    # via pandas
urllib3==2.0.3
    # via requests
uvicorn[standard]==0.22.0
    # via morphology-synthesis (pyproject.toml)
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
